% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{all_variables}
\alias{all_variables}
\title{Names of all variables we can potentially get from netcdfs}
\usage{
all_variables()
}
\description{
Names of all variables we can potentially get from netcdfs
}
\examples{
all_variables()
}
