% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as.GMT}
\alias{as.GMT}
\title{Convert to POSIXct with the right timezone}
\usage{
as.GMT(x)
}
\description{
Convert to POSIXct with the right timezone
}
\examples{
as.GMT('2024-01-01')
as.GMT('2024-01-01 04:00:30')
as.GMT(Sys.time())

}
