% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{date_hours}
\alias{date_hours}
\title{Function for getting all hourly timestamps contained in a vector of date}
\usage{
date_hours(dates)
}
\arguments{
\item{dates}{vector of dates}
}
\description{
Midnight is assigned to the date that is just beginning. Returned times are GMT (as is met Nordic).
}
\examples{
date_hours('2024-01-01')
}
