% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_metno_fc_nn}
\alias{download_metno_fc_nn}
\alias{download_metno_fc_rect}
\title{Download MET Nordic forecast.}
\usage{
download_metno_fc_nn(
  file_out,
  out_dir,
  coords,
  time_start = "all",
  time_end = NULL,
  initialization_time = "latest",
  vars = weather_variables(),
  discard_missing_locations = TRUE
)

download_metno_fc_rect(
  file_out,
  out_dir,
  bbox,
  time_start = "all",
  time_end = NULL,
  initialization_time = "latest",
  vars = weather_variables()
)
}
\arguments{
\item{file_out, out_dir}{file name and directory for saving the results.}

\item{coords}{A data table with columns \code{'lon'}, \code{'lat'}}

\item{time_start}{As in \code{\link[=download_metno_analysis_nn]{download_metno_analysis_nn()}}, but you can also put \code{'all'} for downloading all possible future timestamps.}

\item{time_end}{As in \code{\link[=download_metno_analysis_nn]{download_metno_analysis_nn()}}.}

\item{initialization_time}{Default is \code{'latest'}, which will download the most recent forecast. You can specify downloading older forecasts.}

\item{vars}{character vector of (weather) variables to load, see \code{\link[=weather_variables]{weather_variables()}} to see all available variables.}

\item{discard_missing_locations}{logical. If \code{TRUE} (the default), then coordinates not falling within the met Nordic area are discarded.
If you put this to \code{FALSE}, the function will find a nearest neighbor for all gridcells, even if they are way outside the covered region.}

\item{bbox}{numeric vector of length four giving the boundaries in the order \code{lon_min}, \code{lat_min}, \code{lon_max}, \code{lat_max}.}
}
\description{
The function \code{download_metno_fc_nn} takes a data table of coordinates and downloads the analysis data for the nearest-neighbor gridcells.
The resulting netcdf contains the loaded weather variables and as dimension index a \code{coord_index} which identifies the coordinate.
It moreover has variables \code{lon},\code{lat} (indexed by \code{coord_index}) for retrieving the actual coordinate, as well as
\code{lon_metno}, \code{lat_metno}, the center coordinate for the associate met Nordic gridcell.
Nearest neighbor matching is performed by \code{\link[=get_metno_gridcells_nn]{get_metno_gridcells_nn()}}.
}
\details{
\code{download_metno_fc_rect} downloads the data for a rectangular region specified by \code{bbox}.

Unlike for the analysis-download functions, the loaded data is always saved as single file.
}
\examples{
\dontrun{
# Oslo city hall:
coords = data.table(lat = 59.912,lon = 10.733)

download_metno_fc_nn(file_out = 'oslo_ch_prec_fc.nc',
                     out_dir = tempdir(),
                     coords = coords,
                     vars = 'precipitation_amount')

download_metno_fc_rect(file_out = 'oslo_prec_fc.nc',
                      out_dir = tempdir(),
                      bbox = c(8,58,12,62)
                      vars = 'precipitation_amount')

}

}
