% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{exists_ts_latest}
\alias{exists_ts_latest}
\alias{exists_ts_operational_archive}
\alias{exists_ts_rerun_archive}
\title{In which data streams does a time stamp exist}
\usage{
exists_ts_latest(time, forecast = FALSE)

exists_ts_operational_archive(time)

exists_ts_rerun_archive(time)
}
\arguments{
\item{time}{A vector of times}
}
\value{
logical of the same length as \code{time}
}
\description{
The analyses datasets are organized in three data streams (latest, operational archive, rerun archive), see \href{https://github.com/metno/NWPdocs/wiki/MET-Nordic-dataset}{met Nordic documentation} for details.
These functions check which of the provided timestamps exist in one of the three data streams.
}
\examples{
exists_ts_latest(Sys.Date())
exists_ts_operational_archive(Sys.Date())
exists_ts_rerun_archive(Sys.Date())

exists_ts_latest('2023-05-01')
exists_ts_operational_archive('2023-05-01')
exists_ts_rerun_archive('2023-05-01')

}
