% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fc_fn_and_times}
\alias{get_fc_fn_and_times}
\title{Auxiliary function}
\usage{
get_fc_fn_and_times(time_start, time_end, initialization_time)
}
\arguments{
\item{time_start}{As in \code{\link[=download_metno_analysis_nn]{download_metno_analysis_nn()}}, but you can also put \code{'all'} for downloading all possible future timestamps.}

\item{time_end}{As in \code{\link[=download_metno_analysis_nn]{download_metno_analysis_nn()}}.}

\item{initialization_time}{Default is \code{'latest'}, which will download the most recent forecast. You can specify downloading older forecasts.}
}
\description{
Fetches the forecast file name and the time-indices for netcdf subsetting.
}
