% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_metno_gridcells_rect}
\alias{get_metno_gridcells_rect}
\title{Function for getting met Nordic gridcells in a rectangle area}
\usage{
get_metno_gridcells_rect(bbox)
}
\arguments{
\item{bbox}{numeric vector of length four giving the boundaries in the order \code{lon_min}, \code{lat_min}, \code{lon_max}, \code{lat_max}.}
}
\value{
A data table with columns \code{'lon_metno'}, \code{'lat_metno'} (center coordinates of corresponding met Nordic gridcell),
\code{'x_ind'}, \code{'y_ind'}, \code{'vector_ind'} (indices of the gridcell within the metno netcdfs).
}
\description{
Given a bounding box of the format \code{lon_min}, \code{lat_min}, \code{lon_max}, \code{lat_max}, find all met Nordic gridcells between those boundaries.
}
\examples{
get_metno_gridcells_rect(c(8,58,12,62))

}
\seealso{
get_metno_gridcells_nn
}
