% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gmaps_plot}
\alias{gmaps_plot}
\title{Function for plotting a stars object overlaying google maps}
\usage{
gmaps_plot(
  st,
  bbox = NULL,
  var_name = NULL,
  downsample = "auto100",
  alpha = 0.7,
  tol = 0,
  fix_crs = FALSE,
  map = NULL,
  only_map = FALSE,
  ...
)
}
\arguments{
\item{st}{Data as \code{star}.}

\item{bbox}{Optional bounding box in the form \code{c(lon_min, lat_min, lon_max, lat_max)} (always in lon/lat, no matter the CRS of \code{st}).
If no bounding box is provided, it is derived from \code{st}. Values are approximate, see \code{tol}.}

\item{var_name}{Name of the variable/attribute in \code{st} that you want to plot. Default is to plot the first one.}

\item{downsample}{Should we downsample the data for faster plotting? Default is to downsample to get less than 100.000 pixels.
You can either provide a number (0 means no downsampling, 1 means every second value is kept etc.) or a string of the form
\code{'auto<some number n>'} which will downsample to plot fewer than \code{n}*1000 points, so downsampling rate depends on the size of \code{st}.}

\item{alpha}{Alpha-value for the data. Higher values make the data-plot less opaque (making it harder to see the googlemaps).}

\item{tol}{tolerance value (in lon/lat) for reducing the spatial bounding-box. Only used if \code{bbox = NULL}. Extend of the googlemap is approximate,
so plots frequently look awkward if the full bbox of the data is used, because the map includes areas for which no data is present.}

\item{fix_crs}{This is used to overwrite/set the CRS for netcdfs with wrongly specified CRS: Set this to \code{'metno'}
for plotting MET Nordic data, and to \code{'surf'} for plotting SURF data. Else, \code{st} needs to have the correct CRS}

\item{map}{Optional. You can provide a \code{ggmap} object, then it does not need to be retrieved from googlemaps, since you pay per API-ping.
You can get the map by calling this function with \code{only_map = TRUE}.}

\item{only_map}{Logical. If \code{TRUE}, only the googlemap is returned.}

\item{...}{Passed on to \code{ggmap::get_map()}}
}
\description{
This function take a \code{stars}-object and creates a single map-plot of the data, overlaying a googlemap.
Requires google maps API access. If the \code{stars}-object data for more than 1 time, the first time-slice is plotted.
Returns a gg-object, adjust the color-scale by adding a \code{scale_fill} to it. When calling this many times with the same bbox,
you should provide the map, see parameters \code{map} and \code{only_map}.
}
