% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{open_netcdf_from_url_with_backoff}
\alias{open_netcdf_from_url_with_backoff}
\title{Function to attempt opening a NetCDF file with exponential backoff}
\usage{
open_netcdf_from_url_with_backoff(url, max_retries = 10)
}
\arguments{
\item{url}{URL of netcdf file}

\item{max_retries}{Number of retries. We wait 2^(n-1) secs after the nth try.}
}
\description{
Function to attempt opening a NetCDF file with exponential backoff
}
