% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_analysis_gmap}
\alias{plot_analysis_gmap}
\title{Plot of MET Nordic analysis}
\usage{
plot_analysis_gmap(
  time,
  bbox = c(-180, -90, 180, 90),
  var = "precipitation_amount",
  ...
)
}
\arguments{
\item{time}{Time of the slice you want to plot. See \code{\link[=as.GMT]{as.GMT()}} for details.}

\item{bbox}{Bounding box of the form \code{c(lon_min,lat_min,lon_max,lat_max)} for plotting.}

\item{var}{Name of the variable to plot. See \code{\link[=weather_variables]{weather_variables()}} for what's available.}

\item{...}{passed to \code{\link[=gmaps_plot]{gmaps_plot()}}}
}
\description{
Generate a plot of MET Nordic analysis for a specific timeslice and a rectangular region.
The
}
