% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{read_stars_metno}
\alias{read_stars_metno}
\title{Function for reading downloaded netcdfs as stars}
\usage{
read_stars_metno(...)
}
\arguments{
\item{...}{passed on to \code{\link[stars:read_ncdf]{stars::read_ncdf()}}.}
}
\description{
Corrects the projection, which is not correctly identified by \code{\link[stars:read_stars]{stars::read_stars()}} and \code{\link[stars:read_ncdf]{stars::read_ncdf()}}
}
