% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{separate_fns}
\alias{separate_fns}
\title{Separates filenames into named lists}
\usage{
separate_fns(fns, level)
}
\arguments{
\item{fns}{filenames of download files}

\item{level}{Either \code{'year'}, \code{'month'}, \code{'day'} or \code{'hour'}.}
}
\description{
Auxiliary function used in the download in order to group downloads by year, month, day, or hour.
Takes download filenames and a level to group by, returns a named list.
}
