% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{time_from_fn}
\alias{time_from_fn}
\title{Get a POSIXct from a download-filename}
\usage{
time_from_fn(fns)
}
\arguments{
\item{fns}{Vector of file names of the online netcdfs}
}
\description{
Get a POSIXct from a download-filename
}
