% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{weather_variables}
\alias{weather_variables}
\title{Names of available weather variables}
\usage{
weather_variables()
}
\description{
"precipitation_amount_quantiles" and "precipitation_amount_gt" are excluded - they seem to only be available for some of the analysis products
}
\examples{
weather_variables()
}
