% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which_files_exist_latest}
\alias{which_files_exist_latest}
\title{List files in the METPP-latest catalogue}
\usage{
which_files_exist_latest()
}
\description{
used for checking which files can already be downloaded.
}
