% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_metno_analysis_nn}
\alias{download_metno_analysis_nn}
\title{Download Met Nordic data for selected coordinates}
\usage{
download_metno_analysis_nn(
  file_out,
  out_dir,
  coords,
  time_start,
  time_end = NULL,
  use_rerun = TRUE,
  vars = weather_variables(),
  discard_missing_locations = TRUE
)
}
\arguments{
\item{file_out, out_dir}{file name and directory for saving the results. \code{file_out} needs to end on \code{'.nc'}.}

\item{coords}{A data table with columns \code{'lon'}, \code{'lat'}}

\item{time_start, time_end}{the time stamps you want to extract data for. If \code{time_end} is provided, an hourly sequence from (the lowest entry of) \code{time_start} to \code{time_end} is considered. Alternatively,
a vector of hourly time stamps can be provided as \code{time_start} (if \code{time_end == NULL}). If a vector of dates is provided, all hourly time stamps contained in these dates are used (using UTC timezone, see \code{\link{date_hours}})}

\item{use_rerun}{logical. Decides whether the operational archive or the rerun archive should be preferred (for timestamps for which both is available),
see the \href{https://github.com/metno/NWPdocs/wiki/MET-Nordic-dataset}{met Nordic data documentation} for details.
If \code{TRUE} (the default), the rerun archive is used (which uses one consistent analysis), else the operational archive is used.}

\item{vars}{character vector of (weather) variables to load, see \code{\link{weather_variables}} to see all available variables.}

\item{discard_missing_locations}{logical. If TRUE (the default), then coordinates not falling within the met Nordic area are discarded.}
}
\description{
This function takes a data table of coordinates and downloads the met nordic hourly analysis for these locations and a specified timerange and saves it as netcdf.
The resulting netcdf contains the loaded weather variables and as dimension index a \code{coord_index} which identifies the coordinate.
It moreover has variables \code{lon},\code{lat} (indexed by \code{coord_index}) for retrieving the actual coordinate, as well as
\code{lon_metno}, \code{lat_metno}, the center coordinate for the associate met Nordic gridcell.
Nearest neighbor matching is performed by \code{\link{get_metno_gridcells}}.
}
\examples{
\dontrun{
# Oslo city hall:
data.table(lat = 59.912,lon = 10.733)

download_metno_analysis_nn(file_out = 'oslo_ch_prec_20240101.nc',
                           out_dir = tempdir(),
                           coords = coords,
                           time_start = 2024-01-01,
                           vars = 'precipitation_amount')

}

}
